load_in_8bit_threshold should be a float... Whoops

+++ aiserver.py
@@ -1514,7 +1514,7 @@ def general_startup(override_args=None):
     parser.add_argument("--summarizer_model", action='store', default="philschmid/bart-large-cnn-samsum", help="Huggingface model to use for summarization. Defaults to sshleifer/distilbart-cnn-12-6")
     parser.add_argument("--max_summary_length", action='store', default=75, help="Maximum size for summary to send to image generation")
     parser.add_argument("--multi_story", action='store_true', default=False, help="Allow multi-story mode (experimental)")
-    parser.add_argument("--load_in_8bit_threshold", action='store', default=6, help="set's load_in_8bit_threshold value", type=int)
+    parser.add_argument("--load_in_8bit_threshold", action='store', default=6, help="set's load_in_8bit_threshold value", type=float)
     
     parser.add_argument('-f', action='store', help="option for compatability with colab memory profiles")
     parser.add_argument('-v', '--verbosity', action='count', default=0, help="The default logging level is ERROR or higher. This value increases the amount of logging seen in your screen")
